Extract from Nip19 global funcs into another script
