+++ app/src/main/java/com/vitorpamplona/amethyst/service/Nip19.kt
@@ -2,9 +2,9 @@ package com.vitorpamplona.amethyst.service
 
 import com.vitorpamplona.amethyst.model.toHexKey
 import com.vitorpamplona.amethyst.service.nip19.TlvTypes
+import com.vitorpamplona.amethyst.service.nip19.parseTLV
+import com.vitorpamplona.amethyst.service.nip19.toInt32
 import nostr.postr.bechToBytes
-import java.nio.ByteBuffer
-import java.nio.ByteOrder
 
 class Nip19 {
 
@@ -105,27 +105,4 @@ class Nip19 {
 
         return Return(Type.ADDRESS, "$kind:$author:$d", relay)
     }
-}
-
-fun toInt32(bytes: ByteArray): Int {
-    require(bytes.size == 4) { "length must be 4, got: ${bytes.size}" }
-    return ByteBuffer.wrap(bytes, 0, 4).order(ByteOrder.BIG_ENDIAN).int
-}
-
-fun parseTLV(data: ByteArray): Map<Byte, List<ByteArray>> {
-    val result = mutableMapOf<Byte, MutableList<ByteArray>>()
-    var rest = data
-    while (rest.isNotEmpty()) {
-        val t = rest[0]
-        val l = rest[1]
-        val v = rest.sliceArray(IntRange(2, (2 + l) - 1))
-        rest = rest.sliceArray(IntRange(2 + l, rest.size - 1))
-        if (v.size < l) continue
-
-        if (!result.containsKey(t)) {
-            result[t] = mutableListOf()
-        }
-        result[t]?.add(v)
-    }
-    return result
 }
